% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ee_initialized}
\alias{check_ee_initialized}
\title{Internal: Check that Earth Engine is initialized}
\usage{
check_ee_initialized()
}
\description{
Internal: Check that Earth Engine is initialized
}
\keyword{internal}
