% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_ee_with_progress}
\alias{extract_ee_with_progress}
\title{Extract Earth Engine data with a cli progress bar}
\usage{
extract_ee_with_progress(
  image,
  sf_region,
  scale,
  fun,
  sf,
  quiet = FALSE,
  via = "getInfo",
  ...
)
}
\arguments{
\item{image}{An Earth Engine Image object (from \code{rgee}).}

\item{sf_region}{An \code{sf} object containing regions to extract.}

\item{scale}{Numeric. Scale in meters for extraction.}

\item{fun}{A reducer function, e.g. \code{ee$Reducer$mean()}.}

\item{sf}{Logical. Should the function return an sf object?}

\item{quiet}{Logical. If TRUE, suppress the progress bar (default FALSE).}

\item{via}{Character. Either "getInfo" or "drive".}

\item{...}{arguments of \code{ee_extract} of \code{rgee} packages.}
}
\value{
An \code{sf} or  \code{data.frame} object.
}
\description{
Extract Earth Engine data with a cli progress bar
}
\keyword{internal}
