% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indicators.R
\name{get_indicators}
\alias{get_indicators}
\title{Load and preview indicators from an Excel file}
\usage{
get_indicators(file_path = "files/indicators.xlsx", interactive = FALSE)
}
\arguments{
\item{file_path}{Path to the Excel file (default is "files/indicators.xlsx").}

\item{interactive}{If TRUE, displays a data table in an interactive format (default is FALSE).}
}
\value{
A table with the first rows of the dataset.
}
\description{
This function reads an Excel file containing environmental indicators and displays
the first few rows.
}
