% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics_metada.R
\name{get_metrics_metada}
\alias{get_metrics_metada}
\title{List providers, description, year and link}
\usage{
get_metrics_metada(query = NULL)
}
\arguments{
\item{query}{A string. Default is NULL. List of available providers. For more details, use the \code{get_providers} function.}
}
\value{
A tibble object.
}
\description{
This function allows you to list the list of available providers of the land4health package.
}
\examples{
\donttest{
library(land4health)
get_metrics_metada()
}
}
