% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reducer}
\alias{get_reducer}
\title{Internal: Get an Earth Engine reducer
Returns a reducer object (e.g., \code{ee$Reducer$mean()}) based on a string name.}
\usage{
get_reducer(name)
}
\arguments{
\item{name}{A string: one of \code{"mean"}, \code{"sum"}, \code{"min"}, \code{"max"}, \code{"median"}, \code{"stdDev"} and \code{"first"}}
}
\value{
An Earth Engine reducer object.
}
\description{
Internal: Get an Earth Engine reducer
Returns a reducer object (e.g., \code{ee$Reducer$mean()}) based on a string name.
}
\keyword{internal}
