% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4h_list_metrics.R
\name{l4h_list_metrics}
\alias{l4h_list_metrics}
\title{List available metrics in land4health}
\usage{
l4h_list_metrics(
  category = NULL,
  metric = NULL,
  provider = NULL,
  open_in_browser = TRUE
)
}
\arguments{
\item{category}{Optional string. Filter by thematic category.}

\item{metric}{Optional string. Filter by metric short-name.}

\item{provider}{Optional string. Filter by \code{dataset}.}

\item{open_in_browser}{Logical. Open the URLs in your browser? Default TRUE.}
}
\value{
A tibble (returned invisibly).  When \code{open_in_browser = FALSE}
it prints a compact preview before returning.
}
\description{
List available metrics in land4health
}
