% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{l4h_packages}
\alias{l4h_packages}
\title{List all \emph{land4health} packages}
\usage{
l4h_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{default \code{TRUE}. Includes the "land4health" package name in the
resultant character vector.}
}
\value{
A character vector of package names included in the "land4health" meta-package.
}
\description{
List all \emph{land4health} packages
}
\examples{
l4h_packages()

}
