% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4h_sebal_modis.R
\name{l4h_sebal_modis}
\alias{l4h_sebal_modis}
\title{Download and process evapotranspiration data (SEBAL-MODIS) from Earth Engine}
\usage{
l4h_sebal_modis(
  from,
  to,
  by = "8 days",
  region,
  fun = "mean",
  sf = TRUE,
  force = FALSE,
  ...
)
}
\arguments{
\item{from}{Start date in \code{"YYYY-MM-DD"} format.}

\item{to}{End date in \code{"YYYY-MM-DD"} format.}

\item{by}{Temporal aggregation frequency. Options: \code{"8 days"} (original 8-day composites), \code{"month"} (monthly average or sum), or \code{"annual"} (annual avergae or sumperiod).}

\item{region}{A spatial object defining the region of interest. Accepts \code{sf}, \code{SpatVector}, or \code{ee$FeatureCollection} objects.}

\item{fun}{Aggregation function when \code{by = "month"} or \code{"total"}. Valid values are \code{"mean"} or \code{"sum"}.}

\item{sf}{Logical. Return result as an \code{sf} object? Default is \code{TRUE}.}

\item{force}{Logical. If \code{TRUE}, forces download even if a local file already exists.}

\item{...}{Additional arguments passed to \code{ee_as_stars()} or \code{ee_extract()}.}
}
\value{
A \code{sf} or \code{tibble} object with etp values.
}
\description{
This function accesses the \code{geeSEBAL-MODIS} collection published by the ET-Brasil project,
extracts the \code{etp} band (daily evapotranspiration in mm/day), and allows temporal aggregation
by 8-day images, monthly composites, or the entire period. Optionally, results can be returned
as \code{sf}/\code{stars} objects in R.
}
