% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l4h_to_eedate}
\alias{l4h_to_eedate}
\title{Internal wrapper around rgee::rdate_to_eedate
Mockable with with_mocked_bindings in tests.}
\usage{
l4h_to_eedate(x)
}
\description{
Internal wrapper around rgee::rdate_to_eedate
Mockable with with_mocked_bindings in tests.
}
\keyword{internal}
