% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4h_water_proportion.R
\name{l4h_water_proportion}
\alias{l4h_water_proportion}
\title{Annual Proportion of Water Coverage from MapBiomas Peru}
\usage{
l4h_water_proportion(
  from,
  to,
  region,
  fun = "mean",
  sf = TRUE,
  quiet = FALSE,
  force = FALSE,
  ...
)
}
\arguments{
\item{from}{Integer. Start year (e.g., 1985).}

\item{to}{Integer. End year (e.g., 2022). Must be equal to or greater than \code{from}.}

\item{region}{A spatial object defining the region of interest.
Can be an \code{sf}, \code{sfc} object, or a \code{SpatVector} (from the \pkg{terra} package).}

\item{fun}{Character. Summary function to apply. Values include \code{"mean"}, \code{"sum"},\code{"median"} , etc. Default is \code{"mean"}.}

\item{sf}{Logical. Return result as an \code{sf} object? Default is \code{TRUE}.}

\item{quiet}{Logical. If TRUE, suppress the progress bar (default FALSE).
If \code{FALSE}, returns the Earth Engine \code{ImageCollection}.}

\item{force}{Logical. If \code{TRUE}, skips internal representativity checks on the input geometry.
Defaults to \code{FALSE}.}

\item{...}{arguments of \code{ee_extract} of \code{rgee} packages.}
}
\value{
An object containing annual water coverage. Depending on \code{sf}, it may be an \code{sf} object,
a list, or an Earth Engine \code{ImageCollection} with yearly layers.
}
\description{
The function returns the proportion of each region's area that is covered by surface water for each year.
The values are expressed as a decimal ratio between 0 and 1 (e.g., 0.25 means 25\\% of the area was covered by water).

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# Extract water coverage between 2000 and 2020
l4h_water_proportion(from = 2000, to = 2020, region = my_basin)
}

}
\references{
\itemize{
\item MapBiomas Peru (2021). \emph{Collection 1 – Annual water coverage}.
Mapbiomas project, available on: \url{https://peru.mapbiomas.org/colecciones-de-mapbiomas-peru/}
}
}
