% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{land4health_packages}
\alias{land4health_packages}
\title{List all land4health packages}
\usage{
land4health_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{default \code{TRUE}. Includes the "land4health" package name in the
resultant character vector.}
}
\value{
A character vector of package names included in the "land4health" meta-package.
}
\description{
List all land4health packages
}
\examples{
land4health_packages()

}
