% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_sf}
\alias{split_sf}
\title{Split an sf object into a list of single-row sf objects}
\usage{
split_sf(sf_region)
}
\arguments{
\item{sf_region}{An object of class \code{sf} representing multiple geometries.}
}
\value{
A list of single-row \code{sf} objects.
}
\description{
Split an sf object into a list of single-row sf objects
}
\keyword{internal}
