% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_representativity}
\alias{check_representativity}
\title{Evaluates whether a given polygon covers a minimum number of valid pixels
in a specified Earth Engine image.}
\usage{
check_representativity(region, scale = 30)
}
\arguments{
\item{region}{An \code{sf} polygon object representing the area of interest.}

\item{scale}{Numeric. Pixel resolution in meters (e.g., 30 for Hansen).}
}
\value{
Invisible \code{TRUE} if representative; otherwise \code{FALSE}.
}
\description{
Evaluates whether a given polygon covers a minimum number of valid pixels
in a specified Earth Engine image.
}
\keyword{internal}
