% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4h_sebal_modis.R
\name{l4h_sebal_modis}
\alias{l4h_sebal_modis}
\title{Download and process evapotranspiration data}
\usage{
l4h_sebal_modis(
  from,
  to,
  by = "8 days",
  region,
  fun = "mean",
  sf = TRUE,
  force = FALSE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{from}{Start date in \code{"YYYY-MM-DD"} format.}

\item{to}{End date in \code{"YYYY-MM-DD"} format.}

\item{by}{Temporal aggregation frequency. Options: \code{"8 days"} (original 8-day composites), \code{"month"} (monthly average or sum), or \code{"annual"} (annual avergae or sumperiod).}

\item{region}{A spatial object defining the region of interest. Accepts \code{sf}, \code{SpatVector}, or \code{ee$FeatureCollection} objects.}

\item{fun}{Aggregation function when \code{by = "month"} or \code{"total"}. Valid values are \code{"mean"} or \code{"sum"}.}

\item{sf}{Logical. Return result as an \code{sf} object? Default is \code{TRUE}.}

\item{force}{Logical. If \code{TRUE}, forces download even if a local file already exists.}

\item{quiet}{Logical. If TRUE, suppress the progress bar (default FALSE).}

\item{...}{arguments of \code{ee_extract} of \code{rgee} packages.}
}
\value{
A \code{sf} or \code{tibble} object with etp values.
}
\description{
This function accesses the \code{geeSEBAL-MODIS} collection published by the ET-Brasil project,
extracts the \code{etp} band (daily evapotranspiration in mm/day), and allows temporal aggregation
by 8-day images or monthly or yearly composites period. Optionally, results can be returned
as \code{sf}/\code{tibble} objects in R.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{

library(land4health)
ee_Initialize()

# Define a bounding box region in Ucayali, Peru
region <- st_as_sf(st_sfc(
  st_polygon(list(matrix(c(
    -74.1, -4.4,
    -74.1, -3.7,
    -73.2, -3.7,
    -73.2, -4.4,
    -74.1, -4.4
  ), ncol = 2, byrow = TRUE))),
  crs = 4326
))

# 1. Eight-day composites (8 days)
# 2020-01-01 → 2020-12-31, reducer = "mean"
sebal_8d <- l4h_sebal_modis(
  from   = "2020-01-01",
  to     = "2020-12-31",
  region = region
)

# 2. Monthly means
# Same period, but aggregated to calendar months
sebal_month <- l4h_sebal_modis(
  from   = "2020-01-01",
  to     = "2020-12-31",
  by     = "month",
  region = region
)

# 3. Annual evapotranspiration
# 2015 → 2023, one value per year
sebal_annual <- l4h_sebal_modis(
  from   = 2015,
  to     = 2023,
  by     = "annual",
  fun    = "sum",
  region = region,
  sf     = FALSE
)

}
}
\references{
\itemize{
\item Comini,B., Ruhoff,A., Laipelt,L., Fleischmann,A., Huntington,J.,
Morton,C., Melton,F., Erickson,T., Roberti,D., Souza,V., Biudes,M.,
Machado,N., Santos,C. & Cosio,E. (2023). \emph{geeSEBAL‑MODIS:
Continental‑scale evapotranspiration based on the surface energy
balance for South America.} Preprint. DOI: 10.13140/RG.2.2.17579.11041
\item geeSEBAL‑MODIS v0‑02 dataset. Licensed under the
\emph{Creative Commons Attribution 4.0 International (CC‑BY‑4.0)} license.
}
}
