% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l4h_travel_time.R
\name{l4h_travel_time}
\alias{l4h_travel_time}
\title{Travel Time to Healthcare or Cities (Oxford Dataset)}
\usage{
l4h_travel_time(
  region,
  destination = "cities",
  transport_mode = "all",
  fun = "mean",
  sf = FALSE,
  quiet = FALSE,
  force = FALSE,
  ...
)
}
\arguments{
\item{region}{A spatial object defining the region of interest.
Can be an \code{sf}, \code{sfc} object, or a \code{SpatVector} (from the \pkg{terra} package).}

\item{destination}{Character. Target destination for travel time.
Use \code{"healthcare"} (default) for travel time to the nearest healthcare facility, or
\code{"cities"} for travel time to the nearest populated urban center.}

\item{transport_mode}{Character. Mode of transportation.
Use \code{"all"} (default) for general travel time (mixed modes),
or \code{"walking_only"} for walking-only accessibility (\strong{only valid when \code{destination = "healthcare"}}).}

\item{fun}{Character. Summary function to apply. Values include \code{"mean"}, \code{"sum"},\code{"median"} , etc. Default is \code{"mean"}.}

\item{sf}{Logical. If \code{TRUE}, returns the result as an \code{sf} object. If \code{FALSE},
returns an Earth Engine object. Default is \code{FALSE}.}

\item{quiet}{Logical. If TRUE, suppress the progress bar (default FALSE).}

\item{force}{Logical. If \code{TRUE}, skips the internal representativity check of the input region.
Defaults to \code{FALSE}.}

\item{...}{arguments of \code{ee_extract} of \code{rgee} packages.}
}
\value{
A spatial object containing the computed RAI value for the region in an
\code{sf} or \code{tibble} object.
}
\description{
Retrieves the travel time raster (in minutes) to the nearest healthcare facility
or populated city, based on the Oxford Global Map of Accessibility datasets.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# Travel time to nearest hospital (all transport modes)
l4h_travel_time(my_shape)

# Travel time to hospital (walking only)
l4h_travel_time(my_shape, transport_mode = "walking_only")

# Travel time to cities
l4h_travel_time(my_shape, destination = "cities")

# Mean travel time to cities
l4h_travel_time(my_shape, destination = "cities", reducer = ee$Reducer$mean())
}

}
\references{
\itemize{
\item Weiss, D.J. et al. (2018). \emph{A global map of travel time to cities to assess inequalities in accessibility in 2015.}
Nature, 553(7688), 333–336. DOI: 10.1038/nature25181
\item Weiss, D.J. et al. (2020). \emph{Global maps of travel time to healthcare facilities.}
Nature Medicine, 26, 1835–1838. DOI: 10.1038/s41591-020-1059-1
}
}
