% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land4health-package.R
\docType{package}
\name{land4health-package}
\alias{land4health}
\alias{land4health-package}
\title{land4health: Remote Sensing Metrics for Spatial Health Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate and extract remote sensing metrics for spatial analysis in the field of health. This package offers R users a quick and straightforward way to obtain areal or zonal statistics of key indicators and covariates, ideal for modeling infectious diseases within the framework of spatial epidemiology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://harmonize-tools.github.io/land4health/}
}

}
\author{
\strong{Maintainer}: Yomali Ferreyra \email{yomali.ferreyra@upch.pe} (\href{https://orcid.org/0000-0002-5184-9595}{ORCID})

Authors:
\itemize{
  \item Antony Barja \email{antony.barja@upch.pem} (\href{https://orcid.org/0000-0001-5921-2858}{ORCID})
}

Other contributors:
\itemize{
  \item Diego Villa [contributor]
}

}
\keyword{internal}
